fix: merge jolt-verifier crate into jolt-sdk
