+++ .github/workflows/rust.yml
@@ -95,8 +95,8 @@ jobs:
         working-directory: ./sample_project
         run: cargo run --release
 
-  test-verifier:
-    name: Jolt Verifier Tests
+  test-sdk-verifier:
+    name: Jolt SDK Verifier Tests
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v5
@@ -105,8 +105,8 @@ jobs:
         run: cargo run install-toolchain
       - name: Install nextest
         uses: taiki-e/install-action@nextest
-      - name: Run verifier tests
-        run: cargo nextest run --release -p jolt-verifier
+      - name: Run SDK verifier tests
+        run: cargo nextest run --release -p jolt-sdk --features host
 
   test-tracer:
     name: Jolt Tracer Tests
