+++ Cargo.toml
@@ -29,7 +29,6 @@ members = [
     "jolt-platform",
     "jolt-sdk",
     "jolt-sdk/macros",
-    "jolt-verifier",
     "examples/btreemap/host",
     "examples/btreemap/guest",
     "examples/collatz",
