+++ examples/verifier/guest/src/lib.rs
@@ -1,6 +1,5 @@
-use jolt::{end_cycle_tracking, start_cycle_tracking};
-use jolt_verifier::zkvm::{JoltRV32IM, JoltVerifierPreprocessing, RV32IMJoltProof, Serializable};
-use jolt_verifier::{common, DoryCommitmentScheme, Jolt};
+use jolt::{end_cycle_tracking, start_cycle_tracking, Jolt, PCS};
+use jolt::{JoltRV32IM, JoltVerifierPreprocessing, RV32IMJoltProof, Serializable};
 
 mod fib_io_device_bytes;
 mod fib_proof_bytes;
@@ -13,7 +12,7 @@ mod jolt_verifier_preprocessing_bytes;
 )]
 fn verifier() {
     start_cycle_tracking("preprocessing");
-    let preprocessing: JoltVerifierPreprocessing<ark_bn254::Fr, DoryCommitmentScheme> =
+    let preprocessing: JoltVerifierPreprocessing<ark_bn254::Fr, PCS> =
         JoltVerifierPreprocessing::deserialize_from_bytes(
             jolt_verifier_preprocessing_bytes::VERIFIER_PREPROCESSING_BYTES,
         )
@@ -23,10 +22,8 @@ fn verifier() {
     let proof = RV32IMJoltProof::deserialize_from_bytes(fib_proof_bytes::FIB_PROOF_BYTES).unwrap();
     end_cycle_tracking("proof");
     start_cycle_tracking("device");
-    let device = common::jolt_device::JoltDevice::deserialize_from_bytes(
-        fib_io_device_bytes::FIB_IO_DEVICE_BYTES,
-    )
-    .unwrap();
+    let device =
+        jolt::JoltDevice::deserialize_from_bytes(fib_io_device_bytes::FIB_IO_DEVICE_BYTES).unwrap();
     end_cycle_tracking("device");
     // assert!(device.memory_layout.stack_size > 1024);
     // assert!(proof.trace_length > 0);
