+++ jolt-sdk/src/host_utils.rs
@@ -1,10 +1,14 @@
+#[cfg(feature = "host")]
 pub use ark_bn254::Fr as F;
-pub use common::jolt_device::{MemoryConfig, MemoryLayout};
-pub use jolt_core::field::JoltField;
+#[cfg(feature = "host")]
 pub use jolt_core::host;
+#[cfg(feature = "host")]
+pub use jolt_core::zkvm::dag::proof_serialization::serialize_and_print_size;
+
+pub use common::jolt_device::{JoltDevice, MemoryConfig, MemoryLayout};
+pub use jolt_core::field::JoltField;
 pub use jolt_core::poly::commitment::dory::DoryCommitmentScheme as PCS;
 pub use jolt_core::zkvm::{
-    dag::proof_serialization::serialize_and_print_size, dag::proof_serialization::JoltProof, Jolt,
-    JoltProverPreprocessing, JoltRV32IM, JoltVerifierPreprocessing, RV32IMJoltProof, Serializable,
+    dag::proof_serialization::JoltProof, Jolt, JoltProverPreprocessing, JoltRV32IM,
+    JoltVerifierPreprocessing, RV32IMJoltProof, Serializable,
 };
-pub use tracer::JoltDevice;
