+++ jolt-sdk/tests/fib_io_device_bytes.rs
