+++ jolt-sdk/tests/fib_proof_bytes.rs
