+++ jolt-sdk/tests/fixtures/fib_proof.bin
