+++ jolt-sdk/tests/fixtures/jolt_verifier_preprocessing.dat
