+++ jolt-sdk/tests/gen-fixtures.sh
@@ -5,8 +5,8 @@ echo "Generating fixtures files..."
 cargo run --release -p fibonacci -- --save
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-FIXTURE_DIR=$SCRIPT_DIR/tests/fixtures
-BYTE_RS_DIR=$SCRIPT_DIR/tests
+FIXTURE_DIR=$SCRIPT_DIR/fixtures
+BYTE_RS_DIR=$SCRIPT_DIR
 
 # input/output name pairs: ("input_filename" "output_filename.rs")
 FILES=(
