--- jolt-verifier/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "jolt-verifier"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-jolt-core = { path = "../jolt-core", default-features = false, features = ["parallel"] }
-common = { path = "../common" }
