--- jolt-verifier/src/lib.rs
@@ -1,5 +1,0 @@
-pub use common;
-pub use jolt_core::{
-    field::JoltField, poly::commitment::commitment_scheme::CommitmentScheme,
-    poly::commitment::dory::DoryCommitmentScheme, transcripts::Transcript, zkvm, zkvm::Jolt,
-};
